refactor(consensus): db after storage split
