+++ consensus/src/dag/dag_point_future.rs
@@ -14,9 +14,7 @@ use tycho_util::FastHashMap;
 
 use crate::dag::dag_location::InclusionState;
 use crate::dag::{DagRound, IllFormedReason, ValidateResult, Verifier};
-use crate::effects::{
-    AltFormat, Ctx, DownloadCtx, MempoolStore, RoundCtx, SpawnLimit, TaskResult, ValidateCtx,
-};
+use crate::effects::{AltFormat, Ctx, DownloadCtx, RoundCtx, SpawnLimit, TaskResult, ValidateCtx};
 use crate::engine::NodeConfig;
 use crate::intercom::{DownloadResult, Downloader};
 use crate::models::{
@@ -24,6 +22,7 @@ use crate::models::{
     PointStatusIllFormed, PointStatusNotFound, PointStatusStored, PointStatusStoredRef,
     PointStatusValidated, WeakCert,
 };
+use crate::storage::MempoolStore;
 
 static LIMIT: LazyLock<SpawnLimit> =
     LazyLock::new(|| SpawnLimit::new(NodeConfig::get().max_blocking_tasks.get() as usize));
