+++ consensus/src/effects/mod.rs
@@ -1,6 +1,5 @@
 pub use alt_format::*;
 pub use context::*;
-pub use store::*;
 pub use task::*;
 
 #[macro_use]
@@ -8,5 +7,4 @@ mod macros;
 
 mod alt_format;
 mod context;
-mod store;
 mod task;
