+++ consensus/src/engine/round_task.rs
@@ -8,9 +8,7 @@ use tokio::sync::{oneshot, watch};
 use crate::dag::{
     DagHead, LastOwnPoint, ProduceError, Producer, ValidateResult, Verifier, WeakDagRound,
 };
-use crate::effects::{
-    AltFormat, CollectCtx, Ctx, MempoolStore, RoundCtx, Task, TaskResult, ValidateCtx,
-};
+use crate::effects::{AltFormat, CollectCtx, Ctx, RoundCtx, Task, TaskResult, ValidateCtx};
 use crate::engine::input_buffer::InputBuffer;
 use crate::engine::round_watch::{Consensus, RoundWatch, TopKnownAnchor};
 use crate::intercom::{
@@ -18,6 +16,7 @@ use crate::intercom::{
     Downloader, PeerSchedule, Responder,
 };
 use crate::models::{Cert, Link, Point, PointInfo};
+use crate::storage::MempoolStore;
 
 pub struct RoundTaskState {
     pub peer_schedule: PeerSchedule,
