+++ consensus/src/intercom/core/responder.rs
@@ -7,13 +7,14 @@ use futures_util::{future, FutureExt};
 use tycho_network::{Response, Service, ServiceRequest};
 
 use crate::dag::DagHead;
-use crate::effects::{AltFormat, MempoolStore, RoundCtx};
+use crate::effects::{AltFormat, RoundCtx};
 use crate::intercom::broadcast::Signer;
 use crate::intercom::core::{
     PointByIdResponse, QueryRequest, QueryRequestRaw, QueryRequestTag, QueryResponse,
     SignatureResponse,
 };
 use crate::intercom::{BroadcastFilter, Downloader, Uploader};
+use crate::storage::MempoolStore;
 
 #[derive(Clone, Default)]
 pub struct Responder(Arc<ResponderInner>);
