+++ consensus/src/intercom/dependency/uploader.rs
@@ -4,10 +4,11 @@ use bytes::Bytes;
 use tycho_network::PeerId;
 
 use crate::dag::DagHead;
-use crate::effects::{AltFormat, Cancelled, Ctx, MempoolStore, RoundCtx, SpawnLimit};
+use crate::effects::{AltFormat, Cancelled, Ctx, RoundCtx, SpawnLimit};
 use crate::engine::NodeConfig;
 use crate::intercom::core::PointByIdResponse;
 use crate::models::{PointId, PointStatusStored};
+use crate::storage::MempoolStore;
 
 static LIMIT: LazyLock<SpawnLimit> =
     LazyLock::new(|| SpawnLimit::new(NodeConfig::get().max_upload_tasks.get() as usize));
