+++ consensus/src/lib.rs
@@ -11,7 +11,6 @@ mod storage;
 pub mod test_utils;
 
 pub mod prelude {
-    pub use crate::effects::MempoolAdapterStore;
     pub use crate::engine::lifecycle::{EngineBinding, EngineNetworkArgs, EngineSession};
     pub use crate::engine::round_watch::{RoundWatch, TopKnownAnchor};
     pub use crate::engine::{
@@ -20,4 +19,5 @@ pub mod prelude {
     };
     pub use crate::intercom::InitPeers;
     pub use crate::models::{AnchorData, MempoolOutput, PointInfo};
+    pub use crate::storage::{MempoolAdapterStore, MempoolDb};
 }
