+++ consensus/src/test_utils/dag.rs
@@ -12,7 +12,7 @@ use tycho_network::{Network, OverlayId, PeerId, PrivateOverlay, Router};
 use tycho_util::FastHashMap;
 
 use crate::dag::{AnchorStage, DagRound, ValidateResult, Verifier};
-use crate::effects::{Ctx, EngineCtx, MempoolStore, RoundCtx, TaskTracker, ValidateCtx};
+use crate::effects::{Ctx, EngineCtx, RoundCtx, TaskTracker, ValidateCtx};
 use crate::engine::round_watch::{Consensus, RoundWatch};
 use crate::engine::MempoolConfig;
 use crate::intercom::{Dispatcher, Downloader, InitPeers, PeerSchedule, Responder};
@@ -20,6 +20,7 @@ use crate::models::{
     AnchorStageRole, Cert, Digest, Link, PeerCount, Point, PointData, PointId, Round, Signature,
     Through, UnixTime,
 };
+use crate::storage::MempoolStore;
 
 pub fn make_engine_parts<const PEER_COUNT: usize>(
     peers: &[(PeerId, Arc<KeyPair>); PEER_COUNT],
